로그인 기능에 쓰이는 함수들 세부적으로 분해
