+++ config/config.json
@@ -1,8 +1,8 @@
 {
   "development": {
-    "username": "root",
-    "password": "159753",
-    "database": "moving",
+    "username": "aymrm",
+    "password": "1234",
+    "database": "kdt9",
     "host": "127.0.0.1",
     "dialect": "mysql"
   },
