+++ controller/Cauth.js
@@ -1,8 +1,10 @@
 const {User} = require('../models');
 const crypto = require('crypto');
+const { v4 } = require('uuid');
 const bcrypt = require('bcrypt');
 const constant = require('../common/constant');
 const secret = require('../config/secret');
+const {keylen,digest,maxint,minint} = constant.auth
 
 const dbIdCheck = async (login_id) => {
     const users = await User.findAll({attributes: ["login_id"],where:{login_id}})
@@ -57,15 +59,12 @@ const saltSlicing = (auth_num)=>{
 
 const authCodeIssue = async (user_id)=>{
     const user = await User.findOne({attributes: ["auth","auth_num"],where:{user_id},raw:true})
-    const {keylen,digest,maxint,minint} = constant.auth
     let hash;
     if(user){
         // 정상적인 접근이라면 회원가입을 해서 이미 인증 코드가 들어가 있는 상태에서
         // 로그인으로 접근할때 나오는 코드,
         const {auth,auth_num} = user;
-        const hashSalt = saltSlicing(auth_num + 1);
-        // auth_num에 적힌 숫자에 맞춰 사용할 salt값을 변경
-        hash = await crypto.pbkdf2Sync(auth,hashSalt,auth_num + 1,keylen,digest).toString("base64");
+        hash = await authHashing(auth,auth_num + 1,keylen,digest)
         // 로그인을 새로 했으므로 이전에 사용했던 인증 코드 변경
         const userUpdate = await User.update({
             auth_num : auth_num + 1,
@@ -83,9 +82,34 @@ const authCodeIssue = async (user_id)=>{
     return hash;
 }
 
-const loginCookieRes = async (res,cookieValue)=>{
+const authHashing = async (auth,auth_num) => {
+    const salt = saltSlicing(auth_num);
+    const hash = await crypto.pbkdf2Sync(auth,salt,auth_num,keylen,digest).toString("base64");
+    return hash;
+}
+
+const loginCookieRes = async (login_id,res)=>{
+    const user = await dbIdSearch(login_id);
+    const {user_id,nickname} = user;
+    const auth = await authCodeIssue(user_id);
+    const id = await uuidToString(user_id);
+    const cookieValue = {id,nickname,auth};
     const {loginCookie,cookieSetting} = constant;
     res.cookie(loginCookie,cookieValue,cookieSetting);
+    return nickname;
+}
+
+const signUpConst = async (login_pw)=>{
+    const hash = await pwHashing(login_pw);
+    const uuid = v4();
+    const auth = await authCodeIssue(uuid);
+    const auth_num = crypto.randomInt(maxint)+minint;
+    return {
+        uuid,
+        hash,
+        auth,
+        auth_num,
+    }
 }
 
 const authCheck = async (id,auth)=>{
@@ -93,15 +117,12 @@ const authCheck = async (id,auth)=>{
     const user = await User.findOne({attributes: ["auth","auth_num"],where:{user_id},raw:true})
     if(user){
         // 데이터베이스에 인증 코드가 존재하면 들어오는 코드
-        const {keylen,digest,maxint,minint} = constant.auth
         const {auth:dbAuth,auth_num} = user;
         const hashSalt = saltSlicing(auth_num);
-        console.log("before",hashSalt)
         const hash = await crypto.pbkdf2Sync(dbAuth,hashSalt,auth_num,keylen,digest).toString("base64");
         if(hash === auth){
             const newAuthNum = crypto.randomInt(maxint)+minint;
             const newHashSalt = saltSlicing(newAuthNum);
-            console.log("after",newHashSalt);
             const newAuth = await crypto.pbkdf2Sync(dbAuth,newHashSalt,newAuthNum,keylen,digest).toString("base64");
             const userUpdate = await User.update({
                 auth_num:newAuthNum,
@@ -132,55 +153,6 @@ const authCheckPost = async (req,res)=>{
     res.json({result:false})
 }
 
-const authKakao = async (req,res)=>{
-    try {
-    console.log("query code",req.query.code);
-    const auth = await axios({
-        method: "POST",
-        url: "https://kauth.kakao.com/oauth/token",
-        headers: {
-            "content-type": "application/x-www-form-urlencoded",
-        },
-        data: {
-            grant_type: "authorization_code",
-            client_id: REST_API_KEY,
-            redirect_uri: REDIRECT_URI,
-            code: req.query.code,
-        },
-    });
-    console.log(auth.data);
-        const user = await axios({
-            method: "POST",
-            url: "https://kapi.kakao.com/v2/user/me",
-            headers: {
-                "Authorization": `Bearer ${auth.data.access_token}`,
-                "content-type": "application/x-www-form-urlencoded;charset=utf-8",
-            },
-        });
-        Authorization = `Bearer ${auth.data.access_token}`;
-        target_id = user.data.id;
-        const {profile,birthday,gender} = user.data.kakao_account
-        const data = {
-            id:target_id,
-            authorization:Authorization,
-            nickname:profile.nickname,
-            birth_month:birthday.slice(0,2),
-            birth_day:birthday.slice(2,4),
-            gender,
-        }
-        console.log("data",data)
-        const kakaoIdCheck = await Cauth.dbIdCheck(target_id);
-        let isExist = false;
-        if (kakaoIdCheck){
-            isExist = true;
-        }
-        const {kakaoLoginCookie,cookieSetting} = constant;
-        res.cookie(kakaoLoginCookie,data,cookieSetting);
-        res.redirect('/');
-    } catch (error) {
-        console.log(error)   
-    }
-}
 
 module.exports = {
     dbIdCheck,
@@ -191,7 +163,7 @@ module.exports = {
     stringToUuid,
     authCodeIssue,
     loginCookieRes,
+    signUpConst,
     authCheck,
     authCheckPost,
-    authKakao,
 }
